+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ### Added
 - Add unit and integration tests for Zone model and controller.
 - Add unit and integration tests for Record model and controller.
+- Add Search functionality to search for records by several criteria. Added with tests. 
 
 ### Fixed
 - Fix typo on Releases badge in README.